// src/content/api-reference/floating.d.ts
export declare const floatingPositioning: {
    side: {
        type: {
            type: string;
            definition: string;
        };
        description: string;
    };
    sideOffset: {
        type: string;
        default: string;
        description: string;
    };
    align: {
        type: {
            type: string;
            definition: string;
        };
        description: string;
    };
    alignOffset: {
        type: string;
        default: string;
        description: string;
    };
    avoidCollisions: {
        type: string;
        default: string;
        description: string;
    };
    collisionBoundary: {
        type: {
            type: string;
            definition: string;
        };
        description: string;
    };
    collisionPadding: {
        type: string;
        default: string;
        description: string;
    };
    fitViewport: {
        type: string;
        default: string;
        description: string;
    };
    sameWidth: {
        type: string;
        default: string;
        description: string;
    };
    strategy: {
        type: {
            type: string;
            definition: string;
        };
        default: string;
        description: string;
    };
    overlap: {
        type: string;
        default: string;
        description: string;
    };
};